Conflicts:
	mobile/android/SyncProxyAndroid/src/main/java/com/ford/syncV4/proxy/SyncProxyBase.java
	mobile/android/SyncProxyTester/src/main/assets/CHANGELOG.txt
	mobile/android/SyncProxyTester/src/main/java/com/ford/syncV4/android/activity/SyncProxyTester.java
	src/appMain/life_cycle.cc
	src/appMain/life_cycle.h
	test/components/protocol_handler/include/protocol_handler/protocol_handler_tm_test.h

+++ test/components/json_handler/src/meta_formatter_test_helper.cc
@@ -176,7 +176,7 @@ void CMetaFormatterTestHelper::CompareObjects(
 
   if (NsSmartDeviceLink::NsSmartObjects::SmartType_Array == first.getType()) {
     ASSERT_EQ(NsSmartDeviceLink::NsSmartObjects::SmartType_Array, second.getType());
-    for (int i = 0; i < first.length(); i++) {
+    for (size_t i = 0; i < first.length(); i++) {
       CompareObjects(first.getElement(i), second.getElement(i));
     }
   }
